F/rbac all api

+++ backend/cutie-cuts-app/src/main/java/com/cutie_cuts_app/example/cutie_cuts_app/controller/BookingController.java
@@ -67,7 +67,7 @@ public class BookingController {
         booking.setBarber(barber);
         booking.setDate(request.getDate());
         booking.setTime(request.getTime());
-        booking.setPrice(service.getPrice());
+        booking.setPrice(service.getPrice().doubleValue());
 
         return toResponse(bookingRepository.save(booking));
     }
